+++ src/main/java/com/hp/data/bean/tbox/RemoteControlCmd.java
@@ -5,22 +5,22 @@ import com.hp.data.core.DataEntity;
 @DataEntity(key = "8995")
 public class RemoteControlCmd extends TBoxDownBean{
     private Long eventID;
-    private Long remoteControlType;
+    private Integer remoteControlType;
     private Short acTemperature;
 
     public Long getEventID() {
         return eventID;
     }
 
-    public void setEventID(long eventID) {
+    public void setEventID(Long eventID) {
         this.eventID = eventID;
     }
 
-    public Long getRemoteControlType() {
+    public Integer getRemoteControlType() {
         return remoteControlType;
     }
 
-    public void setRemoteControlType(long remoteControlType) {
+    public void setRemoteControlType(Integer remoteControlType) {
         this.remoteControlType = remoteControlType;
     }
 
@@ -28,7 +28,7 @@ public class RemoteControlCmd extends TBoxDownBean{
         return acTemperature;
     }
 
-    public void setAcTemperature(short acTemperature) {
+    public void setAcTemperature(Short acTemperature) {
         this.acTemperature = acTemperature;
     }
 }
